Refactor component manager and store driver construction
In the component manager, refactor the construction of components to avoid passing unnecessary keyword arguments in Ruby 2.6. If the options are empty, the component is now instantiated without any arguments. Otherwise, the options are passed as keyword arguments.

In the store, refactor the construction of the driver by separating the driver type and driver options. The driver type is fetched from the options, and the driver options are obtained by excluding the driver key from the options.